edge/requirements.txt
numpy>=1.26
scipy>=1.11
PyYAML>=6.0
psycopg2-binary>=2.9
h5py>=3.9
tqdm>=4.66
